Projects with Limit Switches
The .zip works as in one limit switch flipping the direction of the 5th
motor. (Works as of time stamp.) Rest is what I was working on before I
left on Tuesday 2.4.14. Need to set top limit switch to stop arm, pause,
and go back. Bottom needs to stop the motor and have it wait until shot
by driver.